backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
python-multipart==0.0.9
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.143.0
